backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
PyJWT==2.8.0
python-decouple==3.8
django-filter==23.3
pillow==11.3.0
drf-spectacular==0.29.0